refactor: org files into folders
