+++ src/components/shared/SectionHeader.svelte
